Rename form nudge to promo, small css changes
